fix: random unit-test failure with Python >= 3.3
Python 3.3 introduced random in it hashing function to prevent possible
hash collisions attacks. As a result, keys order in a dict are no longer
predictable hence, json serialized dict chances causing request
signature to be different than the expected one. This fix uses
OrderedDict for this specific test case so that the output becomes
predictable again.

Signed-off-by: Jean-Tiare LE BIGOT <jean-tiare.le-bigot@ovh.net>