fix: Don't post data if no arguments were passed.
There have been some changes in the production API that make any POST call without arguments to fail with the following error:

`You provided an input body while none was expected`

This is in particular the case in the `cloud/project/{}/kube/{}/kubeconfig` call.

The `kwargs` arguments to the upstream call is passed as `data` to `raw_call()`. A check for `None` is done to prevent adding a body. This condition is never true as in the case of no arguments, `data` is not `None` but contains an empty dict. In this case, loosy testing for _falsiness_ is better.